Restructure (#2)
* Change file structure, add linter, requirements and gitignore

* added basic config for logging
